+++ src/sections/ShowroomSection/components/ShowroomLink.tsx
@@ -1,0 +1,12 @@
+export const ShowroomLink = () => {
+  return (
+    <div className="box-border caret-transparent pt-[60px] pb-[90px] px-5">
+      <a
+        href="https://farahgorayeb.com/showroom/"
+        className="text-[37.625px] box-border caret-transparent tracking-[-1.88125px] leading-[45.15px] border-b-4 border-solid font-b md:text-[51.2px] md:tracking-[-2.56px] md:leading-[61.44px] md:border-b-[6px] hover:border-transparent"
+      >
+        Learn More About Me
+      </a>
+    </div>
+  );
+};